Rename mapper submodules
